requirements.txt
pydantic==2.7.1
flask==3.0.3
flask-cors
boto3
bson
pyjwt
python-dotenv
Werkzeug
pymongo
python-dotenv
requests
uuid
pandas